Updated user and project import scripts for raven migration.
